Compression module added
