+++ quests/compression/__init__.py
@@ -1,0 +1,2 @@
+
+
